Weird Issue With Pausing
Pausing Seems to be busted as of now , this is most likely due to the timescale being busted , then when switching to new scenes the from the pause menu , the active timescale for the main game scene messes with the rest.

To solve this I would most likely need to make a time manager and give each scene its own seperate time.

Could add this as a prefab maybe?

OR figure a way out to set the new additive scene as the active scene and prevent issues stemming from that